feat: update ProviderRepository interface and test
- Provider와 Product 엔티티 간 연관관계가 존재할 경우 CASCADE 옵션을 통해 관련 엔티티에도 작업을 수행
- Provider 엔티티가 연관관계에서 주인이 아니지만 영속성 전이를 통해 실제 작업을 적용할 수 있도록 함
- in Chapter09

+++ src/test/java/com/springboot/tutorial/ProviderRepositoryTest.java
@@ -4,9 +4,7 @@ import com.springboot.tutorial.repository.ProductRepository;
 import com.springboot.tutorial.repository.ProviderRepository;
 import com.springboot.tutorial.repository.entity.Product;
 import com.springboot.tutorial.repository.entity.Provider;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -24,13 +22,19 @@ public class ProviderRepositoryTest {
 
     @AfterEach
     @DisplayName("테스트가 종료되면 데이터베이스에 저장된 데이터를 삭제한다.")
-    void tearDown() {
+    void tearDown(TestInfo info) {
+        if(info.getTags().contains("SkipAfter"))
+            return;
+
+        // 엔티티에 다른 설정 없이 중간 테이블 삭제 시에는 FK 제약 조건으로 인해 아래와 같은 오류가 발생하게 된다.
+        // Cannot delete or update a parent row: a foreign key constraint fails
         productRepository.deleteAll();
         providerRepository.deleteAll();
     }
 
     @Test
-    @DisplayName("연관관계를 가진 엔티티를 조회하였을 때 Join 연산이 정상적으로 수행되어 결과를 반환한다.")
+    @Tag("SkipAfter") // AfterEach 작업을 수행하지 않도록 태그를 추가한다.
+    @DisplayName("1:N 연관관계를 가진 엔티티를 조회하였을 때 Join 연산이 정상적으로 수행되어 결과를 반환한다.")
     void relationshipTest() {
         Provider provider = new Provider();
         provider.setName("ㅇㅇ물산");
@@ -51,5 +55,42 @@ public class ProviderRepositoryTest {
 
         assertThat(productRepository.findById(1L).orElseThrow(RuntimeException::new)).isNotNull();
         assertThat(productRepository.findById(1L).orElseThrow(RuntimeException::new).getProvider()).isNotNull();
+    }
+
+    @Test
+    @Tag("SkipAfter") // AfterEach 작업을 수행하지 않도록 태그를 추가한다.
+    @DisplayName("영속성 전이를 적용하여 1:N 연관관계를 가진 다른 주인이 아닌 엔티티에 대해서도 CRUD 작업이 정상적으로 수행되어야 한다.")
+    void cascadeTest() {
+        Provider provider = savedProvider("새로운 공급업체");
+
+        Product product1 = savedProduct("상품1", 1000, 1000);
+        Product product2 = savedProduct("상품2", 500, 1500);
+        Product product3 = savedProduct("상품3", 750, 500);
+
+        product1.setProvider(provider);
+        product2.setProvider(provider);
+        product3.setProvider(provider);
+
+        provider.getProduct().addAll(List.of(product1, product2, product3));
+
+        // Provider 엔티티를 대상으로 Insert 작업을 수행하였지만 Product 엔티티에도 Insert 작업이 수행된다.
+        // Provider 엔티티는 1:N 연관관계에서 주인이 아니지만 영속성 전이로 인해 실제 데이터베이스 작업이 수행된다.
+        providerRepository.save(provider);
+    }
+
+    private Provider savedProvider(String name) {
+        Provider provider = new Provider();
+        provider.setName(name);
+
+        return provider;
+    }
+
+    private Product savedProduct(String name, Integer price, Integer stock) {
+        Product product = new Product();
+        product.setName(name);
+        product.setPrice(price);
+        product.setStock(stock);
+
+        return product;
     }
 }
